// frontend/src/backend.d.ts
export interface QuizQuestion {
    question: string;
    correctAnswer: bigint;
    options: Array<string>;
}
export interface CourseModuleContent {
    doStep?: Array<QuizQuestion>;
    show?: string;
    tell: string;
}
export interface CourseWithModules {
    id: bigint;
    reward: bigint;
    title: string;
    availableToAll: boolean;
    description: string;
    assignedMemberTypes: Array<MemberType>;
    category: string;
    modules: Array<CourseModuleWithQuestions>;
}
// ...
export interface CourseModuleWithQuestions {
    id: bigint;
    title: string;
    content: CourseModuleContent;
    description: string;
    questions: Array<QuizQuestion>;
}
// ...
export enum MemberType {
    member = "member",
    trustee = "trustee",
    admin = "admin",
    citizen = "citizen",
    beneficiary = "beneficiary",
    superAdmin = "superAdmin",
    business = "business",
    ambassador = "ambassador",
    partner = "partner"
}